set up cp_nigeria as new app and remove from project files
