+++ app/cp_nigeria/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class CPNigeriaConfig(AppConfig):
+    name = "cp_nigeria"
