+++ app/templates/cp_nigeria/index_cpn.html
@@ -1,4 +1,4 @@
-{% extends 'base.html' %}
+{% extends 'cp_nigeria/base_cpn.html' %}
 {% load static %}
 {% load i18n %}
