+++ app/templates/cp_nigeria/landing_page.html
@@ -1,4 +1,4 @@
-{% extends 'base.html' %}
+{% extends 'cp_nigeria/base_cpn.html' %}
 {% load custom_filters %}
 {% load crispy_forms_tags %}
 {% load static %}
