+++ app/templates/cp_nigeria/navbar_cpn.html
@@ -13,8 +13,8 @@
 
 <nav class="navbar navbar-expand-lg navbar-light bg-light {{ navbar_class }}">
     <div class="container-fluid">
-        <a class="navbar-brand" href="{% url 'project_search' %}">
-            <img src="{% static 'assets/logos/open_plan_logo_horizontal.svg' %}" alt="open_plan Logo">
+        <a class="navbar-brand" href="{% url 'home_cpn' %}">
+            <img src="" alt="CP-Nigeria Logo">
         </a>
         <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
             <span class="navbar-toggler-icon"></span>
