+++ app/templates/cp_nigeria/steps/step_progression.html
@@ -1,8 +1,8 @@
-{% extends 'base.html' %}
+{% extends 'cp_nigeria/base_cpn.html' %}
 {% load i18n %}
 
 {% block navbar %}
-    {% include "navbar.html" with user=user scenario_create=True documentation_url=documentation_url%}
+    {% include "cp_nigeria/navbar_cpn.html" with user=user scenario_create=True documentation_url=documentation_url%}
 {% endblock navbar %}
 
 {% block title %} {% translate "Scenario step" %} {{ step_id }} {% endblock title%}
